fix minor bug in chatglm-onnx
